limit contentlength to u32::MAX - 1
instead of usize::MAX -1.
This is what foo_upnp does, and may prevent problems with some devices.

+++ src/server/streaming_server.rs
@@ -122,7 +122,7 @@ pub fn run_server(
                             rq.remote_addr().unwrap()
                         ));
                         // set transfer encoding chunked unless disabled
-                        let streamsize = Some(usize::MAX - 1);
+                        let streamsize = Some((u32::MAX - 1) as usize);
                         let (tx, rx): (Sender<Vec<f32>>, Receiver<Vec<f32>>) = unbounded();
                         let channel_stream = ChannelStream::new(
                             tx,
@@ -166,7 +166,7 @@ pub fn run_server(
                         ));
                         let response = Response::empty(200)
                             .with_data(channel_stream, streamsize)
-                            .with_chunked_threshold(usize::MAX)
+                            .with_chunked_threshold(u32::MAX as usize)
                             .with_header(cc_hdr)
                             .with_header(ct_hdr)
                             .with_header(tm_hdr)
